examples: use improved SetLog interface
